refactor

+++ utils/progress.py
@@ -1,4 +1,4 @@
 
 
 def calculate_progress_percentage(current_value, total_value=23):
-    return int(100 * (current_value / total_value))
+    return float(100 * (current_value / total_value))
